pdf-chatbot/requirements.txt
streamlit==1.28.0
langchain==0.1.0
PyPDF2==3.0.1
openai==1.6.1
python-dotenv==1.0.0
tiktoken==0.5.2
faiss-cpu==1.7.4